backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0

